refactor: split export tests
